selectedResident...をResident型のselectedResidentにまとめた。getValueAll..の返り値をappStateに定義したresidentsListに変更。他の部分での参照をselectedResident = residentsList[i].givenNameといった感じに変更。main.hにResident構造体を移動させた
