+++ include/main.h
@@ -6,8 +6,22 @@
 
 #include <Arduino.h>
 #include <stack>
-#include "../src/services/api/residents.h"
 #include <vector>
+
+
+//レジデントの情報を識別する構造体
+struct Resident {
+    String residentUid;
+    String familyName;
+    String givenName;
+    
+    // 引数ありコンストラクタ
+    Resident(const String& uid, const String& fam, const String& given)
+        : residentUid(uid), familyName(fam), givenName(given) {}
+    
+    // デフォルトコンストラクタも定義
+    Resident() : residentUid(""), familyName(""), givenName("") {}
+    };
 
 
 // 画面を識別する列挙型
@@ -39,9 +53,7 @@ enum MealTime{
 // アプリケーション全体で使用する状態管理構造体
 struct AppState {
     String selectedUser;     // 選択されたユーザー名
-    std::vector<String> selectedResidentId; // 選択されたレジデントのid
-    std::vector<String> selectedResidentFamilyName;// 選択されたレジデントの苗字
-    std::vector<String> selectedResidentGivenName;// 選択されたレジデントの名前
+    Resident selectedResident; //選択されたレジデントの情報を保持
     Screen currentScreen;    //現在のスクリーン
     std::stack<Screen> screenHistory;//戻る用のスクリーンスタック
     RecordType selectedRecordType; //選択された録音種類
