+++ src/screens/screen_pick_record_type.cpp
@@ -4,8 +4,7 @@ void showRecordTypeScreen(const AppState &state) {
     M5.Lcd.clear();
     M5.Lcd.fillScreen(WHITE);
     M5.Lcd.setTextColor(BLACK, WHITE);
-    String user = state.selectedUser;
-    showHeaderBar("利用者:"+ user);
+    showHeaderBar("利用者:"+ state.selectedResident.givenName);
     M5.Lcd.drawRect(69,  40, 86, 75, BLACK);   // 上段 左
     M5.Lcd.drawRect(165, 40, 86, 75, BLACK);  // 上段 右
     M5.Lcd.drawRect(21,  125, 86, 75, BLACK); // 下段 左
