+++ src/tools/json.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 #include <vector>
-#include "../services/api/residents.h"
+#include <main.h>
 
 String getJsonValue(const String& jsonText, const String& part);
 String getValueInJson(const String& jsonText, const String& part, const String& field);
